% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roverlaps.R
\name{rcovered}
\alias{rcovered}
\title{Return logical of whether a query region intersects subject}
\usage{
rcovered(query, subject, verbose = FALSE)
}
\arguments{
\item{query}{Set of query intervals to evaluate for membership in \code{subject}}

\item{subject}{Set of subject intervals to check \code{query} against}

\item{verbose}{Set the verbosity \code{[FALSE]}}
}
\value{
Logical vector of same length as \code{query}, \code{TRUE} if interval overlaps subject
}
\description{
Determines which members of a query set of regions has any overlap with
a set of subject intervals. The query does not have to be fully contained within
the subject to be counted as \code{TRUE} overlap.
}
\examples{

library(data.table)
o1 <- data.table(seqnames=factor(c("1","2")),start=1,end=5)
o2 <- data.table(seqnames=factor(c("1","1")),start=3,end=5)

## useful for subsetting in one line
o1 <- o1[rcovered(o1,o2)]

## output
#    seqnames start end
# 1:        1     1   5
}
