% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roverlaps.R
\name{rodiff}
\alias{rodiff}
\title{Find the minimum/maximum difference between each element of vector A, compared with set of intervals B of any size}
\usage{
rodiff(query, subject, max = FALSE, sign = 0)
}
\arguments{
\item{query}{Numeric vector to query}

\item{subject}{Numeric vector to query against (start positions of intervals)}

\item{max}{Take the max difference instead of min \code{[FALSE]}}

\item{sign}{If 0, consider values where q > s and s > q, if 1 only consider values where q >= s, if -1 only consider values where s >= q}
}
\value{
Numeric vector of same length as query, holding the differences. Element is NA if no comparison available
}
\description{
For each element of a numeric vector A, this function will find the distance to either the closest or
further element in B. Signs are not considered, as it will internally be computed as absolute value (a - b). If 
a is in b, then distance is 0.
}
\examples{

library(data.table)
set.seed(42)
query <- data.table(seqnames=1, start=sample(100, 3))
subject <- data.table(seqnames=1, start=sample(1000, 100))
subject[, end := start + 10]
#o <- raggediffintervals(query, subject)
}
