% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roverlaps.R
\name{raggeddiff}
\alias{raggeddiff}
\title{Find the minimum/maximum difference between each element of vector A, compared with vector B of any size}
\usage{
raggeddiff(query, subject, max = FALSE, sign = 0)
}
\arguments{
\item{query}{Numeric vector to query}

\item{subject}{Numeric vector to query against}

\item{max}{Take the max difference instead of min \code{[FALSE]}}

\item{sign}{If 0, consider values where q > s and s > q, if 1 only consider values where q >= s, if -1 only consider values where s >= q}
}
\value{
Numeric vector of same length as query, holding the differences. Will return NA if no comparison possible
}
\description{
For each element of a numeric vector A, this function will find the distance to either the closest or
further element in B. Signs are not considered, as it will internally be computed as absolute value (a - b).
}
\examples{

library(data.table)
set.seed(42)
query <- sample(100, 3)
subject <- sample(1000, 100)
#o <- raggediff(query, subject)
}
