% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roverlaps.R
\docType{package}
\name{roverlaps}
\alias{roverlaps}
\alias{roverlaps-package}
\alias{roverlaps}
\title{roverlaps}
\usage{
roverlaps(o1, o2, cores = 1, verbose = FALSE, index_only = FALSE)
}
\arguments{
\item{o1}{Query ranges}

\item{o2}{Subject ranges}

\item{cores}{Maximum number of cores (processes in 1M chunks) [1]}

\item{verbose}{Increase the output to stderr}

\item{index_only}{Return only the indicies (query.id and subject.id)}
}
\value{
data.table ('seqnames', 'start', 'end', 'strand', 'query.id', 'subject.id') of overlaps
}
\description{
Description of your package


}
\author{
Jeremiah Wala <jwala@broadinstitute.org>
}
