% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roverlaps.R
\docType{package}
\name{roverlaps}
\alias{roverlaps}
\alias{cpp_gr2dt}
\alias{roverlaps-package}
\alias{roverlaps}
\title{Converts \code{GRanges} to \code{data.table}}
\usage{
cpp_gr2dt(x)

roverlaps(o1, o2, robust = TRUE)
}
\arguments{
\item{x}{\code{GRanges} to convert}

\item{o1}{Query ranges}

\item{o2}{Subject ranges}

\item{robust}{Perform extra checks (\code{is.unsorted}). \code{[TRUE]}}
}
\value{
data.table of GRanges columns ('seqnames', 'start', 'end', 'strand', 'width') and metadata columns

data.table ('seqnames', 'start', 'end', 'strand', 'query.id', 'subject.id') of overlaps
}
\description{
Converts \code{GRanges} to \code{data.table}
and a field grl.iix which saves the (local) index that that gr was in its corresponding grl item


}
\details{
roverlaps

Description of your package
}
\author{
you <youremail>
}
